backend/requirements.txt
# Core Tools
pip>=23.3
setuptools==80.9.0
wheel>=0.42.0

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# Android Support
adbutils==1.2.1
uiautomator2>=2.16.17,<2.18
apkutils2>=1.0.0

# iOS Support
pymobiledevice3>=2.0.0

# XML Parsing
lxml==5.1.0

# File Upload
python-multipart==0.0.6

# Image Processing
pillow>=10.0.0,<11.0.0

# HTTP Client
requests>=2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
deprecated>=1.2.14
typing-extensions>=4.8.0